requirements.txt
amadeus>=8.0.0
httpx>=0.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
apscheduler>=3.10.0
pyyaml>=6.0
click>=8.0.0
colorama>=0.4.6
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
